% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censcor-package.R
\docType{package}
\name{censcor-package}
\alias{censcor-package}
\title{Bayesian Estimation of Correlations between Censored Variables}
\description{
A set of tools for fully Bayesian estimation of correlations between left and interval censored variables.
}
