\name{censR-package}
\alias{censR-package}
\alias{censR}
\docType{package}
\title{
\packageTitle{censR}
}
\description{
\packageDescription{censR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{censR}
\packageIndices{censR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{censR}

Maintainer: \packageMaintainer{censR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
