% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censR-package.R
\docType{package}
\name{censR-package}
\alias{censR-package}
\alias{censR}
\title{Bayesian Estimation of Correlations between Censored Variables}
\description{
A set of tools for fully Bayesian estimation of correlations between left and interval censored variables.
}
